feat: added exceptions
