+++ src/main/java/project/realtimechatapplication/exception/impl/UnauthorizedMessageDeletionException.java
@@ -1,2 +1,17 @@
-package project.realtimechatapplication.exception.impl;public class UnauthorizedMessageDeletionException {
+package project.realtimechatapplication.exception.impl;
+
+import org.springframework.http.HttpStatus;
+import project.realtimechatapplication.exception.AbstractException;
+
+public class UnauthorizedMessageDeletionException extends AbstractException {
+
+  @Override
+  public int getStatusCode() {
+    return HttpStatus.FORBIDDEN.value();
+  }
+
+  @Override
+  public String getMessage() {
+    return "You are not authorized to delete this message.";
+  }
 }
